Move status module up to module from modman
